+++ lib/http.go
@@ -157,7 +157,7 @@ func init() {
 		for k := range resp.Header {
 			hdr[k] = resp.Header.Get(k)
 		}
-		env.A = nj.Array(nj.Int(int64(resp.StatusCode)), nj.Val(hdr), buf, nj.Val(client.Jar))
+		env.A = nj.Array(nj.Int64(int64(resp.StatusCode)), nj.Val(hdr), buf, nj.Val(client.Jar))
 	}, "$f(options: table) -> array",
 		"\tperform an HTTP request and return { code, headers, body_reader, cookie_jar }",
 		"\t'url' is a mandatory parameter in options, others are optional and pretty self explanatory:",
